+++ lib/opencloud_api_wrapper.dart
@@ -1,2 +1,3 @@
 export 'src/universe.dart' show Universe;
 export 'src/datastore.dart' show Datastore;
+export 'src/ordered_datastore.dart' show OrderedDatastore;