+++ lib/src/urls.dart
@@ -1,2 +1,3 @@
-// ignore: non_constant_identifier_names
+// ignore_for_file: non_constant_identifier_names
 String DATASTORE_API_BASE_ENDPOINT = "https://apis.roblox.com/datastores/v1/universes";
+String ORDERED_DATASTORE_API_BASE_ENDPOINT = "https://apis.roblox.com/ordered-data-stores/v1/universes";